requirements.txt
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-anthropic>=0.3,<0.4
langfuse>=4.0,<5.0
feedparser>=6.0
requests>=2.31
praw>=7.7
exa-py>=1.0
python-dotenv>=1.0
apscheduler>=3.10,<4.0
